Clean up indent
I was struggling a bit to edit the file as it had a mix of tab and
spaces, with what looked like a 2 space indent objective.

Cleaned up the indentation and standardized on 2 space indent. Obviously
no attachment to it, was just try to standardize on what looked like the
original intent